Make the option to use bold/shadow on headings formatters configurable for library users.
