+++ Aztec/Classes/TextKit/Configuration.swift
@@ -1,0 +1,14 @@
+import Foundation
+
+public final class Configuration {
+
+    public static var headersWithBoldTrait = false
+
+    static var defaultBoldFormatter: AttributeFormatter = {
+        if headersWithBoldTrait {
+            return BoldWithShadowForHeadingFormatter()
+        } else {
+            return BoldFormatter()
+        }
+    } ()
+}
